requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
# Only packages your DAG scripts need
boto3
s3fs
apache-airflow-providers-amazon
# tensorflow
# numpy
# pandas
# scikit-learn
# opencv-python
# Pillow
# prometheus-client
